Restructure About section into separate pages
