+++ app/about/experience/page.tsx
@@ -1,0 +1,5 @@
+import { redirect } from "next/navigation"
+
+export default function ExperiencePage() {
+  redirect("/about/experience/kiara")
+}
