+++ locales/ja-JP.json
@@ -31,6 +31,9 @@
     "publications": "出版物・特許",
     "contact": "お問い合わせ"
   },
+  "aboutNav": {
+    "overview": "概要"
+  },
   "experience": {
     "title": "経歴",
     "kiara": {
